some improvements and setting restrictions to delete adjustment

+++ invoices-master/app/Http/Controllers/AdjustmentController.php
@@ -13,21 +13,16 @@ class AdjustmentController extends Controller
     public function index()
     {
         $data = [];
-        $adjustments = [];
         $data['partialView'] = 'adjustment.index';
-        $adjusts = Adjustment::all();
-        foreach ($adjusts as $adjust) { //to check if the user created this object or not
-            if (Auth::id() == $adjust->user_id) {
-                $adjustments[] = $adjust;
-            }
-        }
-        return view('adjustment.base', $data, [
-            'adjustments' => $adjustments]);
+        //to check if the user created this object or not
+        $adjustments = Adjustment::where('user_id', Auth::id())->get();
+        $data['adjustments'] = $adjustments;
+        return view('adjustment.base', $data);
     }
 
     public function init()
     {//create empty record and redirect to edit to have the view displayed then go to the update to save records
-        $adjust = Adjustment::create();
+        $adjust = Adjustment::create(['user_id' => Auth::id()]);
         $id = $adjust->id;
         return redirect(route('adjustments.edit', $id));
     }
@@ -35,13 +30,17 @@ class AdjustmentController extends Controller
     public function edit($id)
     {
         $data = [];
+        $adjustment = Adjustment::where('id', $id)
+            ->where('user_id', Auth::id())
+            ->first();
         $data['partialView'] = 'adjustment.edit';
-        return view('adjustment.base', $data, ['id' => $id]);
+        $data['adjustment'] = $adjustment;
+        return view('adjustment.base', $data);
     }
 
     public function update(Request $request, $id)
     {
-        $adjust = Adjustment::where('id', $id)->update(['name' => $request->name, 'user_id' => Auth::id()]);
+        $adjust = Adjustment::where('id', $id)->update(['name' => $request->name]);
         return response()->json([
             'url' => route('adjustments'),
             'success' => 'record has been saved'
@@ -50,9 +49,18 @@ class AdjustmentController extends Controller
 
     public function destroy($id)
     {
-        Adjustment::destroy($id);
-        return response()->json([
-            'success' => 'Record has been deleted successfully!'
-        ]);
+        $adjustment = Adjustment::where('id', $id)
+            ->where('user_id', Auth::id())
+            ->first();
+        if (count($adjustment->user_invoices)) {
+            return response()->json([
+                'error' => 'Record is related to other recorders in another tables!'
+            ]);
+        } else {
+            Adjustment::destroy($id);
+            return response()->json([
+                'success' => 'Record has been deleted successfully!'
+            ]);
+        }
     }
 }
